pyproject.toml
[project]
name = "gutenbird"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.4.0",
]
